+++ Backend/src/main/java/com/petrosb/EducationalSoftware/attempt/AttemptCreationRequest.java
@@ -1,6 +1,6 @@
 package com.petrosb.EducationalSoftware.attempt;
 
 public record AttemptCreationRequest(
-        Integer score
+        Float score
 ) {
 }
