+++ Backend/src/main/java/com/petrosb/EducationalSoftware/attempt/AttemptUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.petrosb.EducationalSoftware.attempt;
 
 public record AttemptUpdateRequest(
-        Integer score
+        Float score
 ) {
 }
